pyproject.toml
[project]
name = "mimir-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.86",
    "esm>=3.2.1.post1",
    "httpx>=0.28.1",
    "peft>=0.18.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.13",
]
